requirements.txt
streamlit==1.19.0
pandas==1.4.2
folium==0.14.0
streamlit-folium==0.11.1
plotly==5.6.0
python-dotenv==0.21.0
numpy==1.21.5
googlemaps==4.7.3
urllib3==1.26.9
Pillow==9.0.1
requests==2.27.1